Synchronize chain state after notification stream is established. (#4361)
## Motivation

`test_wasm_end_to_end_social_event_streams::remote_net_grpc` is failing
on `devnet-2025-08-15`. The problem seems to be that the relevant block
on the publisher chain arrives _after_ we synchronize that chain
explicitly but _before_ the notification streams from the validators are
established, so we miss that block.

It's difficult to reproduce this issue in another way; locally, the
tests all pass.

## Proposal

Synchronize after establishing the notification stream, for each
validator.

## Test Plan

For me, this fixes the test.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)