Some update to the way ports are assigned. (#4357)
## Motivation

We may want to run with more than 10 validators. A previous PR makes the
work on more than 10 validators impossible.
Here we constrain the product `num_validator * num_shard`.

Fixes #4319 

## Proposal

We assign the ports according to the set constants.
For the test `linera-service/src/exporter/tests.rs` we have a leakage of
the formula. But this is just one test.

## Test Plan

The CI.
The wasm end-to-end tests were run with 20 validators and 1 shard.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

None