Resolve the blob persistence problem (#4356)
## Motivation

Following PR #4247, PR #4278, and PR #3382, we face the scenario where a
contract can create a blob
but not have it available right away. This concerns the access to Blob
Bytecode, Blob ApplicationDescription, and
Blob Data.

This PR addresses those problems.

Fixes #4287 
Fixes #4279 

## Proposal

This PR addresses the problem with the data blob. But it also addresses
the issues for the ApplicationDescription blob
and the contract blobs.

There are three scenarios to cover:
* The blob is created in one block and read in another block.
* The blob is created in one operation and read in the same operation.
* The blob is created in two different operations of the same block.

The first scenario is already working correctly.

The second scenario is addressed with the addition of the test of
`created_blobs` in `ReadBlobContent`. This is done
in the same way as for access to the ApplicationDescription and the
Contract/Service.

The third scenario is addressed by the change in how the
TransactionTracker is built in `block_tracker`.

## Test Plan

The CI and the test in `test_wasm_end_to_end_publish_read_data_blob`,
which considers all 3 scenarios.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

None