Move the `create-and-call` test case from end-to-end to integration test. (#4394)
## Motivation

The end-to-end tests are relatively expensive in CI. So, it is
advantageous to move them to integration tests.

## Proposal

We do the following:
* We remove the `&self` of `ActiveChain` that is not needed.
* We make the function `ActiveChain` public to be used in the
operational test.
* We introduce the function to return `Bytecode` since to instantiate,
we cannot use the compressed bytecode.
* We introduce the test that exercises the `create-and-call` smart
contract.

## Test Plan

The CI.

The PR that introduces the `create-and-call` also addresses the
persistence of the contract and service blobs
that have been created. We tested that this integration test fails if
the correction is removed.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

None.