Remove the caching of `authenticated_signer` / `authenticated_caller_id` in contract calls. (#4374)
## Motivation

The `authenticated_signer` / `authenticated_caller_id` were cached in
the `linera-sdk`.
This is incorrect, since it varies from one call to another.

## Proposal

Caching was removed.

## Test Plan

The CI.

The test of the feature is done with a smart contract `contract-call`.
It is tested in the integration test.
Without this correction, the test fails. The test could not be put at a
lower level like `linera-core` or `linera-execution`
since those do not use the `linera-sdk`.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

None.