Remove test-only `WorkerState.with_key_pair`. (#4401)
## Motivation

This test-only method risks dropping the handles of ongoing tasks with
loaded chain state views, and then loading those chains again.

It's only used in one test, where we assert that the worker updates its
locking block based on a validated block certificate even if it did not
sign to confirm that block. Since this feature mainly matters on the
client side, and we wanted to make sure it can't sign to confirm, we
removed the worker's keys. However, the worker wouldn't sign it anyway,
since it is not in the current round.

## Proposal

Remove the method; assert instead that the worker didn't sign to
confirm.

## Test Plan

Only test code was changed.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)