Properly wait for things that are spawned as bg processes (#4408)
## Motivation

Right now we have a lot of stuff in CI that gets started as a bg
process. However, the issue is that if these things start taking
slightly longer to start up for some reason, it'll make the test fail,
because the test will try to submig requests to something that isn't
ready yet.

## Proposal

Explicitly wait for things to be actually up before proceeding.

## Test Plan

CI

## Release Plan

- Nothing to do / These changes follow the usual release cycle.