Make sure all fields in the GraphQL schema are documented (#4358)
## Motivation

The GraphQL definitions are a little bit inconsistent. Here we correct
them.

Fixes #1308 

## Proposal

The following changes are made:
* Precising the role of the `ChainId`. For example, for `openChain`, a
naive user may think that the input `chainId` is going to be the chain
ID.
* Removing the fact that chains are subscribed to epoch changes from the
admin chain. I think this is not useful as it is an implementation
detail.
* Removing notes like "system API", which are not useful.
* Remove "(or activates)" which is unclear to me and probably obsolete.
* I added the EVM-relevant notes that occurred in a few places.

## Test Plan

The CI should take care of the schema update.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

None