Address a bug with message during instantiation. (#4366)
## Motivation

The scenario of a smart contract emitting a message during instantiation
breaks the `TestValidator` framework.
We address the problem here.

Fixes #3539 

## Proposal

The correction pointed out in the issue is exactly what is needed.
A smart contract has been created that demonstrates that the correction
works and also test
in the end-to-end test, which is nice to check.

## Test Plan

CI.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

None.