Generate operations dynamically (#4345)
## Motivation

Right now we always precalculate the operations for each block. However,
the performance gain we get from it should be minimal, and it limits us
on what we can do.

## Proposal

Generate the operations dynamically instead, giving more flexibility on
what we can do.

## Test Plan

Tested against a local network, works as expected

## Release Plan

- Nothing to do / These changes follow the usual release cycle.