Add single destination per block option to benchmark client (#4331)
## Motivation

Right now the default is for every chain to always send transfers to
every other chain, in a way that in each block we cycle through the
chains.

## Proposal

Add an option so that we continue sending chains to every other chain,
but every block contains transactions to just one of the other chains,
and we cycle through chains per block instead of per transfer.

## Test Plan

Deploy a network and test this mode against it.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.