Use of the `Arc<Box<[u8]>>` for the `CompressedBytecode`. (#4418)
## Motivation

We are using the `Vec<u8>` in the `CompressedBytecode`. But the vector
is never modified so this makes
a copy that is not needed.

## Proposal

The compressed bytecode is now a `Arc<Box<[u8]>>`. An alternative could
have been to use `BlobContent`.
However, we would need to track the `BlobType`, which would be hard.

## Test Plan

The CI.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

None.