Fix spurious error logs in `is_retryable`. (#4432)
## Motivation

`is_retryable` was originally used only for gRPC _streams_, but is now
used in other requests, too. The error-level logs in it are not
appropriate anymore.

In particular, `NotFound` is expected in some cases. And in general, the
logs should never be error-level, because they can be caused by a single
faulty validator.

## Proposal

Fix logs and log levels.

## Test Plan

CI

## Release Plan

- These changes should be backported to the latest `devnet` branch, then
    - be released in a new SDK.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)