Extend reconfig test; sync on `wallet init`; fix blob fetching. (#4437)
## Motivation

We want to make sure the network works fine even after many
reconfigurations, and even if none of the genesis validators are left.

## Proposal

At the end of the reconfiguration end-to-end test, initialize a new
client, request a chain and publish a blob. This test actually failed!
The following two changes make it pass:

* In `wallet init`, synchronize the admin chain, trusting the validators
provided by the faucet (part of
https://github.com/linera-io/linera-protocol/issues/4434).
* Fix a bug in `process_certificates`: Since we are handling a
_certificate_, not a block proposal, we don't need additional proof for
the missing blobs: The certificate itself is proof for them. So we
should only download the blobs themselves, and not try to get another
certificate for them.

## Test Plan

The reconfiguration test was extended.

## Release Plan

- These changes should be backported to the latest `testnet` branch,
then
    - be released in a new SDK.

## Links

- _Partly_ addresses
https://github.com/linera-io/linera-protocol/issues/4434.
- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)