Make `fault_type` part of `LocalValidatorClient`. (#4455)
## Motivation

The `linera-core` unit tests don't run a whole network, and instead wrap
a worker in a mutex inside a clonable `LocalValidatorClient` struct. To
simulate faults, a `FaultType` enum value is stored alongside the worker
inside the mutex.

Since `ChainClient` sometimes cancels tasks that wait for a response
from the validator, and we don't want to cancel the actual worker code
in random places, we `spawn` tasks with the calls to the worker. (Maybe
with the chain state actors that isn't necessary anymore? But that's a
bigger discussion, and such a change could cause new flakiness if we're
missing something.) These spawned tasks have a handle on the mutex with
the worker state and the fault type.

That means the spawned tasks can outlive calls to `ChainClient`
functions in the tests. A `set_fault_type` call after such a function
has returned can still affect how the spawned tasks behave.

I strongly suspect this is what causes the very rare
`test_memory_skipping_proposal` failure in CI, where we observe that a
validator that should have had fault type `DontProcessValidated` ends up
processing a validated block certificate anyway, because _after_ the
client call that had spawned that task, we call `set_fault_type(…
Honest)`.

## Proposal

Put the fault type in the `LocalValidatorClient`, so it is copied before
the task is spawned.

## Test Plan

CI

I will close https://github.com/linera-io/linera-protocol/issues/4422;
let's reopen it if it happens again anyway.

## Release Plan

- These changes should be backported to the testnet and devnet.

## Links

- Fixes https://github.com/linera-io/linera-protocol/issues/4422.
- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)