Allow EVM smart contract to create new EVM contracts. (#4273)
## Motivation

It is very frequent for EVM contracts to create another contract. Thus,
we need to support this feature.

Fixes #3757 

## Proposal

The implementation is surprisingly straightforward:
* Create the blobs with `publish_module`.
* Create the application with `create_application`.
* Return the result according to the format of the EVM.

The last step requires accessing the bytecode of a smart contract. This
is implemented with a direct call.

## Test Plan

In the CI, the test `test_evm_end_to_end_child_subcontract` has been
added and shows the generation of the contract and access to them.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

None.