Make execution state actor borrow txn tracker & resource controller. (#4413)
## Motivation

The transaction tracker is a field in the runtime but also passed into
some execution state methods. In some cases, we take the tracker from
the runtime, send it to the execution state actor, send it back and put
it back into the runtime field. This is confusing and error-prone.

## Proposal

Make the `ExecutionStateActor` a struct that borrows the state, tracker
and resource controller.

## Test Plan

CI

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)