Fix and re-enable `test_finalize_validated`. (#4464)
## Motivation

This test is `#[ignore]`d because it's flaky: If the faulty validators
respond before the other ones start processing the proposal, neither of
them will have the proposal in their chain manager.

Also, `test_sparse_sender_chain` is flaky: If the randomly chosen
validator (the first in the committee, i.e. by public key) is faulty,
processing the notification will fail.

## Proposal

Explicitly handle the proposal in one of the validators to make sure the
test is set up for the intended scenario.

In `test_sparse_sender_chain`, use two validators with no faulty one, so
the selected validator is guaranteed to have the block.

## Test Plan

I ran it 20 times locally and it always passed.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.
- (Backporting would be no problem.)

## Links

- Closes #3860.
- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)