Filter blocks in chains publishing events, introduce sparse chains wrt events (#4466)
## Motivation

If we only listen to a chain because of the events it publishes, we
might not want to download and process all the blocks.

## Proposal

Notify about the events in blocks separately from the notifications
about new blocks. Filter those events client side, preprocess blocks
with events if there are gaps.

(There could still be scenarios in which a client is missing events, and
won't download them, effectively stalling the processing of events on a
chain it owns. This will be taken care of as part of #4255 )

## Test Plan

CI - `test_wasm_end_to_end_social_event_streams` was extended to make
sure that a sparse event publisher chain is created on a client in an
appropriate scenario.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)
- Closes #4261