Update the way that keys are build in views obtained by `linera-views-derive`. (#4419)
## Motivation

It has recently been indicated in PR #4347 that the use `ChainId` in a
message would be a waste of space.
But there are other wastage of key space that can be reduced.

Fixes #2391

## Proposal

We check for the number of entries in views obtained by
`linera-views-derive`. If lower than 256, we convert to `u8`.
It is likely that we need to remove as well the
`linera_views::views::MIN_VIEW_TAG`.

## Test Plan

CI.

## Release Plan

It is definitely breaking the testnet.

## Links

None