Add `epoch` to the `UserChain` and `linera wallet show` (#4458)
## Motivation

The `linera wallet show` presents the data on the chain being tracked.
Other information can be added easily.

Fixes #2958 

## Proposal

The `epoch` is always present in the output of `chain_info_query`. So,
it can be added to the `UserChain`
easily.

Possible extension:
* Adding if a chain is closed or not (it suffices to add a boolean to
`ChainInfo`).

## Test Plan

The readme test case fungible which is the only one with `linera wallet
show` has been

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

None.