[testnet] Clients update validators about missing committees. (#4493) (#4515)
## Motivation

If a validator doesn't have the latest committee and receives a block
signed by that committee, it will fail with `EventsNotFound`.

Also, the validator updater is sending the wrong blocks in some cases
for sparse event publisher chains.

## Proposal

Handle that in the client and update the validator about the admin
chain. Use the correct block heights.

Almost all the credit for this goes to @bart-linera!

## Test Plan

A test was added that reproduced the issue.

## Release Plan

- These changes should be backported to the latest `testnet` branch,
then
    - be released in a new SDK.

## Links

- Closes #4490.
- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)

---------

Co-authored-by: Andreas Fackler <afck@users.noreply.github.com>
Co-authored-by: Bartłomiej Kamiński <bartlomiej.kaminski@linera.io>