Add a test demonstrating how to pass multiple mutations in the `node_service`. (#4539)
## Motivation

It is quite reasonable for users to want to do several operations in
just one block.

## Proposal

The solution is simply to formulate the multiple mutations by using a
construction of the GraphQL queries.

## Test Plan

In order to have multiple operations, an operation `repeated_transfer`
was added to the `NativeFungibleApp`.
The test `test_wasm_end_to_end_same_wallet_fungible` was modified
accordingly.

## Release Plan

The change is fully compatible with the existing codebase.

## Links

None.