`flake.lock`: bump `nixup` (#4553)
## Motivation

In the Nix wrapper, allow setting the Rust toolchain explicitly using
the `RUST_TOOLCHAIN` variable.

This is a bit of a hack until I can get around to reading the
`rust-toolchain.toml` file, but improves the Nix flow for now.

## Proposal

Bump `nixup` to a version that has this hack.

## Test Plan

This is how I build.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)