`linera-base`: expunge `wasmtimer` in favour of `kywasmtime` (#4561)
## Motivation

We ran into a variant of
https://github.com/whizsid/wasmtimer-rs/issues/22, which was drastically
slowing down the operation of `@linera/client` (see #4444 for details).
Additionally, `wasmtimer` doesn't look thread-safe: it uses
[`parking_lot`](https://crates.io/crates/parking_lot), which will panic
on Wasm if the lock ever tries to park.

## Proposal

Replace our dependency on `wasmtimer` with a [Linera
fork](https://github.com/linera-io/kywasmtime/) of
[`kywasmtime`](https://github.com/rom1v/kywasmtime), a package that
exposes the same API but using the browser `clearTimeout` APIs instead
of implementing its own priority queue.

I made a couple of patches to be more helpful for our usage, which are
PR'd upstream as https://github.com/rom1v/kywasmtime/pull/3 and
https://github.com/rom1v/kywasmtime/pull/4.

Since `kywasmtime` isn't published to crates.io, we use it as a git
dependency. This is okay, since `linera-web` is also never published to
crates.io (only to npm), and the SDK crates that we do publish don't
rely on it.

## Test Plan

Manually tested.

## Release Plan

- These changes should be backported to the latest `testnet` branch,
then
    - be released in a new SDK.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)
- Fixes #4444.