Restructure `linera-web` directory (#4577)
## Motivation

We are unhappy with the directory structure of `linera-web`, which looks
too much like an SDK crate.

## Proposal

Rename it to plain old `web` to indicate that it's not (published as) a
crate, with the npm package structure (`@linera/client` and
`@linera/signer`) reproduced inside.

## Test Plan

CI.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)