Implement the syncing with all validators functionality. (#4558)
## Motivation

In operations, it is useful to be able to sync with all validators in
one single operation.

Fixes #4488 

## Proposal

The implementation is straightforward.

## Test Plan

The CI.
I did test with README.md kind of files and added a `linera
sync-all-validators` function. And it does access to the validators
without error.
I did not test that the syncing does happen, though.

I created issue #4535 that shows some idea for improving the syncing
with validators.

## Release Plan

Fully compatible with existing DevNet / TestNet.

## Links

None.