Add some useful client commands (#4570)
## Motivation

It is hard to find out the ID of the admin chain, or to check the
contents of blocks using the CLI client.

## Proposal

This PR adds some useful commands to the client:
`show-network-description` (which contains the admin chain ID), `chain
show-chain-description` and `chain show-block`.

## Test Plan

Tested manually.

## Release Plan

TBD (could be backported to the testnet and devnet branches, but it's
not strictly necessary).

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)