Add feature to use jemalloc as memory allocator (#4571)
## Motivation

jemalloc is more performant than the system allocator (and was even the
Rust default for a while)

## Proposal

Add a feature that when enabled will use jemalloc as the allocator. We
might decide to have this feature on by default eventually, but for now
it'll be used for memory profiling, which is coming in the following
PRs.

## Test Plan

Deployed a network with this featured turned on, saw a slight
performance improvement, and tested with the profiling code.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.