Continuous memory profiling (#4572)
For debugging production memory issues, ideally we'll have continuous
memory profiling always on.

This exposes a couple endpoints:
* One that generates a pprof and returns it once the endpoint is queried
* One that generates a flamegraph and returns it once the endpoint is
queried

The former will be used by Alloy on the deployment side, to generate
profiles every 30 seconds or so. It shows no visible performance
regression.

Deployed a network with this and the `linera-infra` changes, saw the
memory flamegraphs on Pyroscope, and managed to curl the flamegraph
endpoint and get a memory profile svg returned to me.

![Screenshot 2025-09-06 at
02.27.50.png](https://app.graphite.dev/user-attachments/assets/f5c2407e-af25-4b8c-b43a-9e56258a9373.png)

- Nothing to do / These changes follow the usual release cycle.