Decrease block cache size, increase execution state cache size (#4544)
## Motivation

After a while, while benchmarking, some of our shards start OOMing.
After inspecting memory profiles, it seems that the culprit is the
`ValueCache` for blocks specifically.

## Proposal

Lower the size of the cache so that it fills up and start evicting
entries before making the VM run out of memory.

## Test Plan

Deploy a network with this code, benchmark it for a while and make sure
that memory goes flat

## Release Plan

- Nothing to do / These changes follow the usual release cycle.