Change the argument of `handle_query`. (#4585)
## Motivation

The `call_application` and `instantiate` are taking a reference, but in
contrast to those functions the `handle_query``
is taking a value which is immediate serialized. There is no reason for
that.



## Proposal

Make the change of having `handle_query` take a reference.
Also change `create_data_blob`.

## Test Plan

The CI.

## Release Plan

This is definitely breaking the smart contracts, though not the
validator code.

## Links

None.