Implement the fees for the events. (#4592)
## Motivation

Events costs need to be accounted for.

Fixes #4259

## Proposal

Since events are implemented as blobs, it makes sense to have their
costs identical to those of blobs.
In that way, we avoid having some economic imbalance.

## Test Plan

The CI.

## Release Plan

This does change the costs for validators and so it is a breaking
change.

## Links

None.