Download missing blobs when retrying outgoing messages (#4597)
## Motivation

We were seeing `Failed to retry outgoing messages` errors on Testnet
Conway due to blobs missing on the client side.

## Proposal

If blobs are missing during execution of this part of the code, download
them and retry.

The change is purely client-side, so it is safe to be backported to the
testnet branch.

## Test Plan

Manual testing showed that the errors disappear.

## Release Plan

- These changes should be backported to the latest `testnet` branch

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)