Add monitoring to node service (#4607)
## Motivation

Currently, if we start a node service we get no metrics or memory
profiling.

## Proposal

Add support for metrics and memory profiling from node service. Also fix
the fact that it doesn't terminate with Ctrl+C (the faucet had the same
thing, so I fixed it there too).

## Test Plan

Start a node service locally, see that memory profile is enabled and
working, and that killing it with Ctrl+C actually kills it

## Release Plan

- Nothing to do / These changes follow the usual release cycle.