Implement `BLOCK_CACHE_SIZE` / `EXECUTION_STATE_CACHE_SIZE` as input parameters (#4621)
## Motivation

Those two variables are constants in the program while varying values
could help.

## Proposal

Add it as parameter inputs.
A creation function has been introduced, which leads to a simplification
of the code.

## Test Plan

The CI.

## Release Plan

The goal is to backport it to testnet conway. But let's first merge in
in the main.

## Links

None.