Adding Tempo for distributed tracing (#4556)
Distributed tracing is a great way to debug different types of issues,
including for example latency issues. So this is something we definitely
want in general, and probably want by default in production as well.

Implement Distributed Tracing using Grafana Tempo. As it is a Grafana
product, it integrates well with it, which is great for us. The
visualizations also seem to be decent.

Deployed a network with this code and the `linera-infra` portion of
this, and everything works as expected, and I can see the latency
breakdowns (I got a really high latency outlier example):

![Screenshot 2025-09-16 at
13.48.59.png](https://app.graphite.dev/user-attachments/assets/98f49272-d04a-4b7e-aa83-c04f90ec7347.png)

I also chose this because it shows we might be waiting in the chain
worker channel's queue for a while here 🤔 which might be worth
investigating, which I'll do next.

- Nothing to do / These changes follow the usual release cycle.