Add more instrumentation (#4555)
Now that we have distributed tracing (after
https://github.com/linera-io/linera-protocol/pull/4556), we need more
instrumentation so we have data about more functions in the breakdowns.

Instrument more functions with `telemetry_only` so that we don't get
spammed in our logs, but the spans still get sent to Tempo.

Tested this with https://github.com/linera-io/linera-protocol/pull/4556,
saw the spans properly show in the breakdowns.

- Nothing to do / These changes follow the usual release cycle.