Allow multi-read in `CollectionView` (#4609)
## Motivation

The reading of multiple keys in `CollectionView` does not work. This is
unfortunate since the underlying
container of `CollectionView` is `async_lock::RwLock` which allows
multiple read or one write. We `CollectionView`
should have the same semantics.

Fixes #4588 

## Proposal

The following steps are made:
* The `ReadGuardedView` is now an enum. The `Loaded` entry is when it is
present in updates. The `NotLoaded` when accessed from storage.
* The `NotLoaded` contains some lock so as to avoid a write to be
created and lead to incoherent state.
* As a consequence, when doing the `try_load_entry` of a view on
storage, the updates is not changed. But that is fine, since it is
anyway in the LRU cache.
* A `try_load_entries` has been introduced since we can now read
multiple keys at once.
* The code of `fn entries` has been simplifed to use those functions.
* The `try_load_entries` no longer take values since only references are
needed.
* The `Clone` requirement have been alleviated in `CollectionView` /
`ReentrantCollectionView`.


## Test Plan

The CI.
The unit tests in the code should be sufficient for the functionality.

## Release Plan

It addresses an issue that showed up in the GameOfLife. So, moving it to
TestNet seems like.a good idea.

## Links

None.