Streamline the use of GraphQLMutationRoot. (#4538)
## Motivation

The `GraphQLMutationRoot` is used for some smart contracts like
`non-fungible` when it is not needed.

## Proposal

There are two ways to implement the trait `MutationRoot`:
* One is to use the function `mutation_root` of the trait
`GraphQLMutationRoot` implemented by the derive macro. For example,
`counter`.
* Implement the trait explicitly. For example, `non-fungible`.

Therefore, two changes are made:
* When the second strategy is used, we should not implement the
`GraphQLMutationRoot` (3 cases removed).
* When the trait is being used, we formulate it as `GraphQLMutationRoot
as _` which is clearer (and examples are supposed to help).

## Test Plan

The CI.

## Release Plan

It does not break anything.

## Links

None.