Move some tests to a separate directory. (#4406)
## Motivation

Several tests have emerged that are unit test in nature but require a
full application to be built.
Currently, the source code of those tests is in `examples` which makes
them sit alongside more
functional applications that are of more general interest. 

## Proposal

The tests are moved to the directory `linera-sdk/tests/fixtures`. They
are not part of the strusture of the
cargo crate.

The advantage is that those smart contracts are no longer compiled again
and again (ScyllaDb, DynamoDb,
StorageService, RemoteNet), and so we gain on CI here. The cost of
compiling for tests in `--release` is about
2-3 minutes per such instance.

The cost of running `linera-sdk/tests/fixtures` is 7 minutes, so
overall, we have a speed gain from this work.

## Test Plan

The CI.
The same tests are being tested. The same tests are being applied, just
differently ordered.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

None.