[Backport] Add a limit for the number of received_log entries in a chain info response (#4643)
## Motivation

The huge amount of entries in the `received_log` seems to be making it
impossible to sync the old GoL chain on Testnet Conway.

## Proposal

Limit the number of returned entries in a single response. Client should
request more entries with an increased offset until the number of
returned entries is less than the limit (indicating that there are no
more entries).

## Test Plan

CI; Manual testing on the testnet will show whether this fixes the
syncing issue

## Release Plan

- These changes should be:
    - released in a new SDK,
    - released in a validator hotfix.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)

---------

Co-authored-by: Mathieu Baudet <1105398+ma2bd@users.noreply.github.com>